finish switch to float64 and fixed build

+++ randvar/utils.go
@@ -3,8 +3,8 @@ package randvar
 import "errors"
 
 // VerifyDist verifies if distribution sum is 1
-func VerifyDist(event string, distribution map[string]float32) error {
-	var sum float32 = 0.0
+func VerifyDist(event string, distribution map[string]float64) error {
+	var sum float64 = 0.0
 	for _, d := range distribution {
 		sum += d
 	}
